@@ -9,18 +9,14 @@ android {
             storePassword 'android'
         }
     }
-    compileSdkVersion rootProject.ext.compileSdkVersion
-    buildToolsVersion rootProject.ext.buildToolsVersion
-    //compileSdkVersion 21
-    //buildToolsVersion '21.1.1'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.rodrigo.lock.app"
-        minSdkVersion rootProject.ext.minSdkVersion
-        targetSdkVersion rootProject.ext.targetSdkVersion
-//        minSdkVersion 14
-//        targetSdkVersion 21
-        versionCode 19
-        versionName "0.2.7"
+        minSdkVersion 15
+        targetSdkVersion 28
+        versionCode 20
+        versionName "0.2.8"
         //renderscriptTargetApi 21
         //renderscriptSupportMode true
         multiDexEnabled true
