@@ -3,10 +3,16 @@
 buildscript {
     repositories {
         jcenter()
+        maven {
+            url 'https://maven.google.com/'
+            name 'Google'
+        }
+        google()
+        maven { url 'https://maven.fabric.io/public' }
     }
     dependencies {
        // classpath 'com.android.tools.build:gradle:0.12.+'
-        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath 'com.android.tools.build:gradle:3.6.3'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -15,22 +21,26 @@ buildscript {
 
 allprojects {
     repositories {
-        google()
         jcenter()
+        mavenCentral()
+        maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
+        maven { url 'http://dl.bintray.com/amulyakhare/maven' }
+        maven { url "https://jitpack.io" }
+        google()
     }
 }
 
 
 ext {
     // Sdk and tools
-    minSdkVersion = 14
-    targetSdkVersion = 22
-    compileSdkVersion = 23
-    buildToolsVersion = '26.0.2'
+    //minSdkVersion = 14
+    //targetSdkVersion = 28
+    //compileSdkVersion = 28
+    // buildToolsVersion = '28.0.3'
 
     // App dependencies
-    supportLibraryVersion = '24.+'
-    guavaVersion = '18.0'
+    //supportLibraryVersion = '28.+'
+    //guavaVersion = '18.0'
     //junitVersion = '4.12'
    // mockitoVersion = '1.10.19'
     //powerMockito = '1.6.2'
@@ -38,7 +48,7 @@ ext {
     //runnerVersion = '0.4.1'
     //rulesVersion = '0.4.1'
     //espressoVersion = '2.2.1'
-    rxjavaVersion = '1.1.8'
-    rxandroidVersion = '1.2.1'
+    //rxjavaVersion = '1.1.8'
+    //rxandroidVersion = '1.2.1'
     //sqlbriteVersion = '0.7.0'
 }
