@@ -1,9 +1,9 @@
 package com.rodrigo.lock.app.mvp.createVault;
 
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 
 import com.rodrigo.lock.app.R;
 import com.rodrigo.lock.app.utils.ActivityUtils;
