@@ -1,7 +1,7 @@
 package com.rodrigo.lock.app.mvp.createVault;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.text.TextUtils;
 
 import com.rodrigo.lock.app.data.Clases.Vault;
