@@ -2,18 +2,17 @@ package com.rodrigo.lock.app.mvp.listVaults;
 
 import android.content.Intent;
 import android.net.Uri;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import com.google.android.material.navigation.NavigationView;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.Toast;
 
-import com.google.android.gms.wallet.Wallet;
 import com.rodrigo.lock.app.mvp.backup.BackupActivity;
 import com.rodrigo.lock.app.mvp.preguntasFrecuentes.PreguntasFrecuentes;
 import com.rodrigo.lock.app.utils.Injection;
@@ -72,11 +71,12 @@ public class VaultsActivity extends AppCompatActivity {
                 tasksFragment,
                 Injection.provideSchedulerProvider());
 
-
+/*
         Date actual = new Date();
         if (actual.compareTo(fechaHabilitaPagos) < 0){
             hideDonar();
-        }
+        }*/
+        hideDonar();
     }
 
 
