@@ -2,13 +2,13 @@ package com.rodrigo.lock.app.mvp.listVaults;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.Fragment;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.widget.SwipeRefreshLayout;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.fragment.app.Fragment;
+import androidx.core.content.ContextCompat;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.MotionEvent;
@@ -210,7 +210,7 @@ public class VaultsFragment extends Fragment implements VaultsContract.View  {
                 //.setCaptureTouchEventOutsidePrompt(true)
                 .setPrimaryText(getResources().getString(R.string.lista_bobeda_vacio_titulo))
                 .setSecondaryText(getResources().getString(R.string.lista_bobeda_vacio_descripcion))
-                .setOnHidePromptListener(new MaterialTapTargetPrompt.OnHidePromptListener(){
+                /*.setOnHidePromptListener(new MaterialTapTargetPrompt.OnHidePromptListener(){
                     @Override
                     public void onHidePrompt(MotionEvent event, boolean tappedTarget){
                         //Do something such as storing a value so that this prompt is never shown again
@@ -223,7 +223,7 @@ public class VaultsFragment extends Fragment implements VaultsContract.View  {
                     public void onHidePromptComplete(){
 
                     }
-                })
+                })*/
                 .show();
     }
 
