@@ -3,7 +3,7 @@ package com.rodrigo.lock.app.mvp.migrar;
 import android.content.Context;
 import android.content.Intent;
 import android.os.AsyncTask;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 
 import com.rodrigo.lock.app.Constants;
